[feat💡]: saves the son's file inside a folder.
I've created a single file pattern for son registration, that concists in grouping all information related to them inside a single one folder named with the son's name. Inside that file it can be stored multiple data, such as, bank account data, company data, and many others related to them.

+++ src/familymanagersystem/forms/RegistrationMainScreenSonRegistration.java
@@ -179,46 +179,58 @@ public class RegistrationMainScreenSonRegistration extends javax.swing.JPanel {
                 JOptionPane.showMessageDialog(null, "You Need To Fill All The Fields", "Please, Fill Out All Fields", 0);
             } else {
                 //holds all information related to the son
-                File sonInformation = new File(RegistrationMainScreenInfoFields.defaultFamilyFilePath + "\\" + RegistrationMainScreenInfoFields.familyLastname + "\\" + sonName);
+                File sonInformation = new File(RegistrationMainScreenInfoFields.defaultFamilyFilePath + RegistrationMainScreenInfoFields.familyLastname + "\\" + sonName + "\\about.txt");
 
-                //checks if there's already a registered son with the same given name
-                if (sonInformation.exists()) {
-                    //shows an error message
-                    JOptionPane.showMessageDialog(null, "<html>You already registered a </html>" + sonName, "Registration Failed", 1);
-                    //clears all the form fields
-                    jSonNameField.setText("");
-                    jBornDateField.setText("");
-                } else {
+                //main directory for all the files related to registrated son
+                File info = new File(RegistrationMainScreenInfoFields.defaultFamilyFilePath + RegistrationMainScreenInfoFields.familyLastname + "\\" + sonName);
 
-                    //creates the file inside the given family folder
-                    try {
-                        if (sonInformation.createNewFile()) {
+                //creates the main directory
+                info.mkdir();
 
-                            //setts all values to the son class
-                            try ( //creates a writer for the file
-                                    FileWriter sonInformationFileWriter = new FileWriter(sonInformation)) {
-                                //setts all values to the son class
-                                Filhos.setBornDate(bornDate);
-                                Filhos.setGender("Male");
-                                Filhos.setName(sonName);
-                                //writes all info in the file
-                                sonInformationFileWriter.write("Name:" + Filhos.getName() + "\n");
-                                sonInformationFileWriter.write("Gender:" + Filhos.getGender() + "\n");
-                                sonInformationFileWriter.write("Born Date:" + Filhos.getBornDate() + "\n");
-                                //closes the writer object
-                            }
-
-                            //shows a success message
-                            JOptionPane.showMessageDialog(null, sonName + " Was Successfully Registered !", "Registration Complete", 2);
-
-                            //clears the input fields
-                            jSonNameField.setText("");
-                            jBornDateField.setText("");
-                        }
-                    } catch (IOException e) {
+                //checks if the directory was created
+                if (info.exists()) {
+                    //checks if there's already a registered son with the same given name
+                    if (sonInformation.exists()) {
                         //shows an error message
-                        JOptionPane.showMessageDialog(null, "<html>There's been a error while <br> <strong>trying to save the info</strong> about<br>the son</html>", "Registration Failed", 1);
+                        JOptionPane.showMessageDialog(null, "<html>You already registered a </html>" + sonName, "Registration Failed", 1);
+                        //clears all the form fields
+                        jSonNameField.setText("");
+                        jBornDateField.setText("");
+                    } else {
+
+                        //creates the file inside the given family folder
+                        try {
+                            if (sonInformation.createNewFile()) {
+
+                                //setts all values to the son class
+                                try ( //creates a writer for the file
+                                         FileWriter sonInformationFileWriter = new FileWriter(sonInformation)) {
+                                    //setts all values to the son class
+                                    Filhos.setBornDate(bornDate);
+                                    Filhos.setGender("Male");
+                                    Filhos.setName(sonName);
+                                    //writes all info in the file
+                                    sonInformationFileWriter.write("Name:" + Filhos.getName() + "\n");
+                                    sonInformationFileWriter.write("Gender:" + Filhos.getGender() + "\n");
+                                    sonInformationFileWriter.write("Born Date:" + Filhos.getBornDate());
+                                    //closes the writer object
+                                    sonInformationFileWriter.close();
+                                }
+
+                                //shows a success message
+                                JOptionPane.showMessageDialog(null, sonName + " Was Successfully Registered !", "Registration Complete", 2);
+
+                                //clears the input fields
+                                jSonNameField.setText("");
+                                jBornDateField.setText("");
+                            }
+                        } catch (IOException e) {
+                            //shows an error message
+                            JOptionPane.showMessageDialog(null, "<html>There's been a error while <br> <strong>trying to save the info</strong> about<br>the son</html>", "Registration Failed", 1);
+                        }
                     }
+                } else {
+                    JOptionPane.showMessageDialog(null, "Couldn't find " + sonName + "info folder", "Error 404: Folder Not Found", 0);
                 }
             }
         }
